Rename src/infra to src/utils
Resolves #98
